orig-monorepo/sms-api-gateway/requirements.txt
absl-py==2.3.0
access==1.1.9
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autograd==1.8.0
babel==2.17.0
beautifulsoup4==4.13.4
bigraph-schema @ file:///Users/alexanderpatrie/Desktop/repos/ecoli/bigraph-schema
biopython==1.85
bleach==6.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.0
click==8.2.1
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
component-contribution==0.6.0
contourpy==1.3.2
cryptography==45.0.3
cvxpy==1.6.5
cycler==0.12.1
cython==3.1.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
dnspython==2.7.0
duckdb==1.3.0
ecos==2.0.14
equilibrator-api==0.6.0
equilibrator-assets==0.6.0
equilibrator-cache==0.6.1
esda==2.7.0
ete3==3.1.3
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fiona==1.10.1
fire==0.7.0
flexcache==0.3
flexparser==0.4
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
gcsfs==2025.5.1
genecoli @ file:///Users/alexanderpatrie/Desktop/repos/ecoli/genEcoli
geopandas==1.0.1
giddy==2.3.6
google-api-core==2.24.2
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
inequality==1.1.1
iniconfig==2.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==9.2.0
ipython-genutils==0.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
iteround==1.0.4
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
lazy-loader==0.4
levenshtein==0.27.1
libpysal==4.13.0
line-profiler==4.2.0
llvmlite==0.44.0
mapclassify==2.9.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mgwr==2.2.1
mistune==3.1.3
momepy==0.10.0
mpmath==1.3.0
multidict==6.4.4
narwhals==1.41.0
nbclassic==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
openbabel-wheel==3.1.1.22
opencv-python-headless==4.11.0.86
orjson==3.10.18
ortools==9.11.4210
osqp==1.0.4
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
path==16.16.0
patsy==1.0.1
periodictable==2.0.2
pexpect==4.9.0
pillow==11.2.1
pint==0.24.4
pip==25.1.1
platformdirs==4.3.8
pluggy==1.6.0
pointpats==2.5.1
polars==1.30.0
pooch==1.8.2
process-bigraph @ file:///Users/alexanderpatrie/Desktop/repos/ecoli/process-bigraph
prometheus-client==0.22.0
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.26.1
psutil==7.0.0
ptyprocess==0.7.0
pulp==3.1.1
pure-eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pymongo==4.13.0
pymunk==6.11.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyqt5==5.15.11
pyqt5-qt5==5.15.17
pyqt5-sip==12.17.0
pysal==25.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
quantecon==0.8.0
rapidfuzz==3.13.0
rasterio==1.4.3
rasterstats==0.20.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
rustworkx==0.16.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
scs==3.2.7.post2
seaborn==0.13.2
segregation==2.5.2
send2trash==1.8.3
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spaghetti==1.7.6
spglm==1.1.0
spint==1.0.7
splot==1.1.7
spopt==0.6.1
spreg==1.8.3
sqlalchemy==2.0.41
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
stochastic-arrow==1.1.0
swiglpk==5.0.12
sympy==1.14.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.26
tinycss2==1.4.0
tobler==0.12.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uncertainties==3.2.3
unum==4.2.1
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
-e file:///Users/alexanderpatrie/Desktop/repos/ecoli/vEcoli
vivarium-core==1.6.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
yarl==1.20.0
